+++ postgres/src/lib.rs
@@ -90,11 +90,9 @@ pub mod compat {
 pub mod dev {
     //! traits for extending functionalities through external crate
 
-    pub use crate::client::ClientBorrowMut;
+    pub use crate::client::{ClientBorrow, ClientBorrowMut};
     pub use crate::copy::r#Copy;
     pub use crate::driver::codec::{Response, encode::Encode, response::IntoResponse};
-    pub use crate::prepare::Prepare;
-    pub use crate::query::Query;
 }
 
 use core::{future::Future, pin::Pin, sync::atomic::AtomicUsize};